Converted to use the Antlr grammar found here - https://github.com/jorendorff/rust-grammar - tests pass, and generally it works.
Currently we munge the grammar rules somewhat, to add explicit token names for tokens which were defined inline;
this also fixes some issues in the original grammar by introducing a stable order for token definition (for
example, not needing to define '<<' as '<' '<').

Next tasks:  Move our munging to a fork of the original grammar;  address cases from the rust lang sources that
do not parse;  introduce named sub-rules that allow the parser to easily identify things such as macro and function
invocations (which the original grammar just bundles into "expr'), so that features such as find usages are
implementable.

Some remnants of javacc-related code persist, particularly in tests which were adapted to run using the new
parser.